Move OnlineAccounts into its own process
Create the com.canonical.OnlineAccountsUI D-Bus service, whose task is to show
the Online Accounts UI when clients request it.
The SystemSettings plugin for Online Accounts becomes just one of these clients.

The rationale for the change is that it has been decided by design that
multiple instances of Online Accounts can be running at the same time (because
the Online Accounts UI will be modal to the application which invoked it), yet
they need to coordinate in order to avoid the creation of the same account in
two different places. Using a single process with multiple windows and exposing
a D-Bus API seems to be the best approach and is also flexible enough to
accomodate for most of the design changes that might pop up. It's also a proved
approach, used already by signon-ui (in fact, most of the code is taken from
there).