Fix QQmlExpression's invalid context warning
The bug was caused because we are destroying the page while the Ubuntu UI
toolkit is still playing the fading out animation for the dialog which prompts
for the removal confirmation.
The fix consists of emitting the closed() signal (and therefore proceed with
the removal of the account) only after the dialog animation has been completed.
This is a sequential animation, whose last step is setting the "visible"
property to false.